Add support for full config-files.
Introduces a bit of "wrapper logic" to the document parsing so that bare
waves sections are added to a dummy object to match the structure of a
full config-file. Additionally, an arena selection dropdown box is added
to the top of the right-hand side. This allows for pasting in an entire
config-file and just selecting which arena to analyze.